Moved cmap benchmarks to its own file
